XRates v1.0

+++ XRates/XRates/DAL/EF/Entities/Rate.cs
@@ -6,9 +6,7 @@ namespace XRates.DAL.EF.Entities
     {
         public int Id { get; set; }
         public DateTime Date { get; set; }
-        public decimal Value { get; set; }
+        public double Value { get; set; }
         public string Code { get; set; }
-
-        //public Country Country { get; set; }
     }
 }
